Added base list code comments
