+++ example/lib/pages/base-list/paginated-base-list.page.dart
@@ -53,6 +53,7 @@ class PaginatedBaseListPage extends StatelessWidget {
                 padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 20),
                 builder: (_, account) => AccountTile(account: account),
                 separatorBuilder: (_) => const SizedBox(height: 16),
+                emptyStateBuilder: (context) => const Text("Empty state"),
               ),
             ),
             BaseListPaginator<Accounts>(
