+++ lib/src/base-list/config/public.dart
@@ -1,4 +1,4 @@
 export './base-list-config.data.dart';
-export './base-list-config.widget.dart';
+export '../widgets/base-list-config.widget.dart';
 export './base-list-data-filter.provider.dart';
 export './base-list-mode.enum.dart';
